+++ common/src/main/java/smartin/miapi/client/modelrework/BakedMiapiModel.java
@@ -2,14 +2,23 @@ package smartin.miapi.client.modelrework;
 
 import com.redpxnda.nucleus.util.Color;
 import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.render.OverlayTexture;
+import net.minecraft.client.render.TexturedRenderLayers;
 import net.minecraft.client.render.VertexConsumer;
 import net.minecraft.client.render.VertexConsumerProvider;
+import net.minecraft.client.render.entity.feature.ArmorFeatureRenderer;
 import net.minecraft.client.render.model.BakedModel;
 import net.minecraft.client.render.model.json.ModelOverrideList;
 import net.minecraft.client.render.model.json.ModelTransformationMode;
+import net.minecraft.client.texture.Sprite;
+import net.minecraft.client.texture.SpriteAtlasTexture;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.item.ArmorItem;
+import net.minecraft.item.ArmorMaterial;
+import net.minecraft.item.ArmorMaterials;
 import net.minecraft.item.ItemStack;
+import net.minecraft.item.trim.ArmorTrim;
 import net.minecraft.util.math.Direction;
 import net.minecraft.util.math.random.Random;
 import org.joml.Matrix4f;
@@ -27,6 +36,7 @@ public class BakedMiapiModel implements MiapiModel {
     ModelHolder modelHolder;
     Random random = Random.create();
     float[] colors;
+    private final SpriteAtlasTexture armorTrimsAtlas = MinecraftClient.getInstance().getBakedModelManager().getAtlas(TexturedRenderLayers.ARMOR_TRIMS_ATLAS_TEXTURE);
 
     public BakedMiapiModel(ModelHolder holder, ItemModule.ModuleInstance instance, ItemStack stack) {
         modelHolder = holder;
@@ -56,12 +66,29 @@ public class BakedMiapiModel implements MiapiModel {
                     .forEach(bakedQuad -> consumer.quad(matrices.peek(), bakedQuad, colors[0], colors[1], colors[2], light, overlay));
         }
 
+        if (stack.getItem() instanceof ArmorItem armorItem) {
+            ArmorTrim.getTrim(entity.getWorld().getRegistryManager(), stack).ifPresent((trim) -> {
+                this.renderTrim(armorItem.getMaterial(), matrices, vertexConsumers, light, trim, model, false);
+            });
+        }
+
         MinecraftClient.getInstance().world.getProfiler().pop();
         MinecraftClient.getInstance().world.getProfiler().pop();
         matrices.pop();
     }
 
-    public record ModelHolder(BakedModel model, Matrix4f matrix4f, ColorProvider colorProvider) {
+    private void renderTrim(ArmorMaterial material, MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, ArmorTrim trim, BakedModel model, boolean leggings) {
+        Sprite sprite = this.armorTrimsAtlas.getSprite(leggings ? trim.getLeggingsModelId(material) : trim.getGenericModelId(material));
+        VertexConsumer vertexConsumer = sprite.getTextureSpecificVertexConsumer(vertexConsumers.getBuffer(TexturedRenderLayers.getArmorTrims()));
+
+        for (Direction direction : Direction.values()) {
+            model.getQuads(null, direction, random)
+                    .forEach(bakedQuad -> vertexConsumer.quad(matrices.peek(), bakedQuad, colors[0], colors[1], colors[2], light, OverlayTexture.DEFAULT_UV));
+        }
+    }
+
+    public record ModelHolder(BakedModel model, Matrix4f matrix4f, ColorProvider colorProvider,
+                              TrimRenderer.TrimMode trimMode) {
 
     }
 }