+++ common/src/main/java/smartin/miapi/item/FakeEnchantment.java
@@ -10,24 +10,33 @@ import java.util.Map;
 
 public class FakeEnchantment {
 
-    public static Map<Enchantment, List<EnchantmentTransformer>> enchantmentListMap = new HashMap<>();
+    public static Map<Enchantment, List<SpecificEnchantmentTransformer>> enchantmentListMap = new HashMap<>();
+
+    public static List<EnchantmentTransformer> enchantmentTransformers = new ArrayList<>();
 
     public static int getFakeLevel(Enchantment enchantment, ItemStack stack, int level) {
         if (enchantmentListMap.get(enchantment) != null) {
-            for (EnchantmentTransformer transformer : enchantmentListMap.get(enchantment)) {
+            for (SpecificEnchantmentTransformer transformer : enchantmentListMap.get(enchantment)) {
                 level = transformer.level(stack, level);
             }
+        }
+        for(EnchantmentTransformer transformer:enchantmentTransformers){
+            level = transformer.level(enchantment,stack,level);
         }
         return level;
     }
 
-    public static void addTransformer(Enchantment enchantment, EnchantmentTransformer transformer) {
-        List<EnchantmentTransformer> transformers = enchantmentListMap.getOrDefault(enchantment, new ArrayList<>());
+    public static void addTransformer(Enchantment enchantment, SpecificEnchantmentTransformer transformer) {
+        List<SpecificEnchantmentTransformer> transformers = enchantmentListMap.getOrDefault(enchantment, new ArrayList<>());
         transformers.add(transformer);
         enchantmentListMap.put(enchantment, transformers);
     }
 
     public interface EnchantmentTransformer {
+        int level(Enchantment enchantment, ItemStack stack, int level);
+    }
+
+    public interface SpecificEnchantmentTransformer {
         int level(ItemStack stack, int level);
     }
 }