+++ CHANGELOG.md
@@ -2,3 +2,7 @@
 - added new expiremental Alt-render
 - Altrenderer should be fully compatible with Iris, Oculus and Optifine
 - This Altrenderer can be toggled in the miapi.json config  the default Option is AUTO, use DEFAULT_RENDERER or FALLBACK_RENDERER in case any issues are encounted. Please report any encountered issues with the Alt Renderer
+- blocked Archery AutofireHook on Modular Crossbows to prevent crash
+- made craftingEnchants apply on Smithing as well
+- fixed crash when rapidly modifying items in workbench
+- hopefully fixed crossbows unable to shoot under some conditions